Make nodes static inner classes
