+++ src/main/java/minijava/ast/ClassDeclaration.java
@@ -1,8 +1,6 @@
 package minijava.ast;
 
 import java.util.List;
-import minijava.ast.classmember.ClassMember;
-import minijava.ast.visitors.ClassDeclarationVisitor;
 
 public class ClassDeclaration<TRef> {
   public final String name;
@@ -13,7 +11,12 @@ public class ClassDeclaration<TRef> {
     this.members = members;
   }
 
-  public <TRet> TRet acceptVisitor(ClassDeclarationVisitor<TRef, TRet> visitor) {
+  public <TRet> TRet acceptVisitor(Visitor<TRef, TRet> visitor) {
     return visitor.visitClassDeclaration(name, members);
+  }
+
+  public interface Visitor<TRef, TReturn> {
+
+    TReturn visitClassDeclaration(String name, List<ClassMember<TRef>> members);
   }
 }
