+++ src/main/java/minijava/ast/Program.java
@@ -1,7 +1,6 @@
 package minijava.ast;
 
 import java.util.List;
-import minijava.ast.visitors.ProgramVisitor;
 
 public class Program<TRef> {
   public final List<ClassDeclaration<TRef>> declarations;
@@ -10,7 +9,12 @@ public class Program<TRef> {
     this.declarations = declarations;
   }
 
-  public <TRet> TRet acceptVisitor(ProgramVisitor<TRef, TRet> visitor) {
+  public <TRet> TRet acceptVisitor(Visitor<TRef, TRet> visitor) {
     return visitor.visitProgram(declarations);
+  }
+
+  public interface Visitor<TRef, TReturn> {
+
+    TReturn visitProgram(List<ClassDeclaration<TRef>> declarations);
   }
 }
