--- src/main/java/minijava/ast/classmember/ClassMember.java
@@ -1,7 +1,0 @@
-package minijava.ast.classmember;
-
-import minijava.ast.visitors.ClassMemberVisitor;
-
-public interface ClassMember<TRef> {
-  <TRet> TRet acceptVisitor(ClassMemberVisitor<TRef, TRet> visitor);
-}
