--- src/main/java/minijava/ast/expression/BinOp.java
@@ -1,18 +1,0 @@
-package minijava.ast.expression;
-
-public enum BinOp {
-  ASSIGN,
-  PLUS,
-  MINUS,
-  MULTIPLY,
-  DIVIDE,
-  MODULO,
-  OR,
-  AND,
-  EQ,
-  NEQ,
-  LT,
-  LEQ,
-  GT,
-  GEQ
-}
