--- src/main/java/minijava/ast/expression/Expression.java
@@ -1,7 +1,0 @@
-package minijava.ast.expression;
-
-import minijava.ast.visitors.ExpressionVisitor;
-
-public interface Expression<TRef> {
-  <TRet> TRet acceptVisitor(ExpressionVisitor<TRef, TRet> visitor);
-}
