--- src/main/java/minijava/ast/expression/UnOp.java
@@ -1,6 +1,0 @@
-package minijava.ast.expression;
-
-public enum UnOp {
-  NOT,
-  NEGATE,
-}
