--- src/main/java/minijava/ast/statement/BlockStatement.java
@@ -1,7 +1,0 @@
-package minijava.ast.statement;
-
-import minijava.ast.visitors.BlockStatementVisitor;
-
-public interface BlockStatement<TRef> {
-  <TRet> TRet acceptVisitor(BlockStatementVisitor<TRef, TRet> visitor);
-}
