--- src/main/java/minijava/ast/type/BooleanType.java
@@ -1,10 +1,0 @@
-package minijava.ast.type;
-
-import minijava.ast.visitors.TypeVisitor;
-
-public class BooleanType<TRef> implements Type<TRef> {
-  @Override
-  public <TRet> TRet acceptVisitor(TypeVisitor<TRef, TRet> visitor) {
-    return visitor.visitBoolean();
-  }
-}
