--- src/main/java/minijava/ast/type/IntType.java
@@ -1,10 +1,0 @@
-package minijava.ast.type;
-
-import minijava.ast.visitors.TypeVisitor;
-
-public class IntType<TRef> implements Type<TRef> {
-  @Override
-  public <TRet> TRet acceptVisitor(TypeVisitor<TRef, TRet> visitor) {
-    return visitor.visitInt();
-  }
-}
