--- src/main/java/minijava/ast/type/Type.java
@@ -1,7 +1,0 @@
-package minijava.ast.type;
-
-import minijava.ast.visitors.TypeVisitor;
-
-public interface Type<TRef> {
-  <TRet> TRet acceptVisitor(TypeVisitor<TRef, TRet> visitor);
-}
