--- src/main/java/minijava/ast/type/VoidType.java
@@ -1,10 +1,0 @@
-package minijava.ast.type;
-
-import minijava.ast.visitors.TypeVisitor;
-
-public class VoidType<TRef> implements Type<TRef> {
-  @Override
-  public <TRet> TRet acceptVisitor(TypeVisitor<TRef, TRet> visitor) {
-    return visitor.visitVoid();
-  }
-}
