--- src/main/java/minijava/ast/visitors/BlockStatementVisitor.java
@@ -1,9 +1,0 @@
-package minijava.ast.visitors;
-
-import minijava.ast.expression.Expression;
-import minijava.ast.type.Type;
-
-public interface BlockStatementVisitor<TRef, TRet> extends StatementVisitor<TRef, TRet> {
-
-  TRet visitLocalVariableDeclarationStatement(Type<TRef> type, String name, Expression<TRef> rhs);
-}
