--- src/main/java/minijava/ast/visitors/ProgramVisitor.java
@@ -1,9 +1,0 @@
-package minijava.ast.visitors;
-
-import java.util.List;
-import minijava.ast.ClassDeclaration;
-
-public interface ProgramVisitor<TRef, TReturn> {
-
-  TReturn visitProgram(List<ClassDeclaration<TRef>> declarations);
-}
