--- src/main/java/minijava/ast/visitors/TypeVisitor.java
@@ -1,16 +1,0 @@
-package minijava.ast.visitors;
-
-import minijava.ast.type.Type;
-
-public interface TypeVisitor<TRef, TReturn> {
-
-  TReturn visitArray(Type<TRef> elementType);
-
-  TReturn visitClass(TRef classRef);
-
-  TReturn visitVoid();
-
-  TReturn visitBoolean();
-
-  TReturn visitInt();
-}
